apps/steel-defect/requirements.txt
dash
plotly
dash-bootstrap-components
gunicorn
numpy
pandas
scipy
tensorflow-cpu==1.15
Pillow
scikit-learn